Added obs mode specifier to JWST API script
This allows the user to specify an observation mode (e.g. "IFU", "MOS",
"IMAGE" etc., and also adds an optional argument to the input parser.

In short: With this request, the user can specify whether they want e.g.
MIRI Images or MIRI IFU data, to spare their internet connection and
their time. If the user does *not* want to think about this, they don't
need to do anything.